# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/patient/get-patient/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Patient Management"
                ],
                "summary": "Get Patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patientmodel.GetPatientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
        "/patient/list-patient": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "List Employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Patient Management"
                ],
                "summary": "List Employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "name"
                        ],
                        "type": "string",
                        "description": "Sort field. Default to name,asc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ignore pagination and return all data",
                        "name": "paging_ignored",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patientmodel.ListPatientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
        "/patient/medical": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Upsert Medical Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Patient Management"
                ],
                "summary": "Upsert Medical Record",
                "parameters": [
                    {
                        "description": "Upsert medical record Request",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patient.UpsertMedicalRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiEmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
        "/patient/medical/history": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get Medical History",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Patient Management"
                ],
                "summary": "Get Medical History",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Ignore pagination and return all data",
                        "name": "paging_ignored",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From date",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "To date",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Created by, doctor ID or nurse ID",
                        "name": "created_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patientmodel.GetMedicalHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
        "/patient/medical/history/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get Medical History Detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SA - Patient Management"
                ],
                "summary": "Get Medical History Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical History ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patientmodel.GetMedicalHistoryDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/basemodel.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "basemodel.ApiEmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "basemodel.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gcommon.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "gcommon.ErrorDetail": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "patient.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "$ref": "#/definitions/patient.User"
                },
                "date_of_birth": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "updated_by": {
                    "$ref": "#/definitions/patient.User"
                }
            }
        },
...
        "patient.UpsertMedicalRecordRequest": {
            "type": "object",
            "properties": {
                "diagnosis": {
                    "type": "string"
                },
                "doctor_notes": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "medical_end_date": {
                    "type": "integer"
                },
                "patient_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "patientmodel.GetMedicalHistoryDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/patientmodel.GetMedicalHistoryDetailResponseData"
                }
            }
        },
...
        "patientmodel.GetMedicalHistoryResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "medical_history": {
                    "$ref": "#/definitions/patientmodel.MedicalHistoryData"
                }
            }
        },
        "patientmodel.GetPatientResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "patient": {
                    "$ref": "#/definitions/patient.Patient"
                }
            }
        },
...
        "patientmodel.ListPatientResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/patientmodel.ListPatientResponseData"
                }
            }
        },
...
    }
}